style: prevented lecture-description field to be broken into parts
When a description was broken into parts, due to the column, layout, sometimes the content of a description was split into parts. As a consequence, the overall design of those parts was off and seemed broken. This is why, the splitting of a description field, not the container, has been eliminated.