fix: fixed duplicate key error
This error occurred as the id of an empty-lecture-list-item was set to '' and this item was set for each element of the Array with the size of 30 elements. Hence, all of the items had the same key. For that reason the id property's value was set to a random number and the array is no longer filled with the same value but instead mapped over, which calls the dummyBasic function for each one.