fix: fixed filtering of lectures in LectureSearch
Previously the lectures were filtered inside a client component that used a somewhat global state used for the rendering of lectures and filtering. Now that the state is handled by the lectureProvider its value includes the lectures-state but also the initial-lectures.

For that reason the if condition at the beginning of the search-handler was modified to check for valid initialLectures but also the filtering check it self was modified to use the preFiltered variable instead of the lectures variable to filter by type.