fix!: fixed visual-bug on safari where type background was cut off
The reason for this visual bug was that the the upper part of the circle that shows the type of a lecture was cut off and displayed at the end of the previous column. As a consequence the lecture-item was wrapped inside of a parent div that includes the label in its dimensions by setting a padding at the top. By doing so, the label does no longer wrap to another column as it is inside of the new parent div.