feat: updated /lectures redirection route
When a user is logged in and a study and a semester were previously selected, then the user will be redirected to that previous page. On the other hand when the user is not logged in or does not have this data stored, then he will be redirected to the corresponding page to set the corresponding preferences.