refactor: updated LectureList props to accept routeParams
This component now no longer needs the SessionData to determinate the full route of lecture-item, as it gets the study and semester route params passed by its parent. This can be done as the study-selection is validated in the root-page of the lecture-display process (/[semester]/page) and thus the case of an invalid study may not occur.