feat: updated /[semester]/page preference-store of semester
When a user opens the /[semester]/page then the underlying study is saved if its valid and the same goes for the semester once its validity is checked it is being stored in the users' preferences, if the user is logged in. If the validity of one of the parameters is invalid then the user is redirected to the corresponding page to make a selection.