fix: fixed loading-lecture-data & display
Previously the dummyLectures variable inside of the loading page was empty. Thus the array is first filled with the same values and then mapped over to ensure that each element is different.

Also the returned LectureList container has the property isPending set to true and an array with an length of 5 to display a skeleton for the semester-selection.