feat: added responsiveness to lecture-type
The lecture-type icon is now also responsive, hence the isPending property had to be added and extracted from the components properties. Also the display of the lectures type has been replaced with the responsive text component. In addition, the background is set by random, when isPending is set to true, so that the page feels more dynamic.